src/backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Trading and backtesting
backtrader==1.9.78.123
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Data processing and analysis
ta==0.10.2
TA-Lib==0.4.28
scikit-learn==1.3.2

# Real-time data and WebSockets
websockets==12.0
redis==5.0.1
celery==5.3.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0

# HTTP clients for broker APIs
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional broker integrations
# ib-insync==0.9.86  # Interactive Brokers
# alpaca-trade-api==3.1.1  # Alpaca